% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractFastqcBaseNs}
\alias{extractFastqcBaseNs}
\title{Extract FastQC}
\usage{
extractFastqcBaseNs(all.qc)
}
\arguments{
\item{all.qc}{return value from readFastqcZips()}
}
\value{
numeric matrix with entries in rows and positions along sequences in columns
}
\description{
Returns the N count along the sequences per entry in a set of zip archives generated by FastQC.
To be used after readFastqcZips().
When reads are long, FastQC can average N count over a given range, e.g. over two successive bases, such as "11-12".
In such a case, because the ranges can be different for different fastq files in \code{all.qc}, this function will set N count to the first base of the range in the ouput, here "11", the other, here "12", will get NA.
}
\author{
Timothee Flutre [cre,aut]
}

