% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractFastqcNreads}
\alias{extractFastqcNreads}
\title{Return the number of sequences per entry in a set of zip archives generated by FastQC.}
\usage{
extractFastqcNreads(all.qc)
}
\arguments{
\item{all.qc}{return value from readFastqcZips()}
}
\value{
numeric vector
}
\description{
To be used after readFastqcZips().
}
\author{
Timothee Flutre [cre,aut]
}

