% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractFastqcQuals}
\alias{extractFastqcQuals}
\title{Extract FastQC}
\usage{
extractFastqcQuals(all.qc, perc = FALSE, nreads = NULL)
}
\arguments{
\item{all.qc}{return value from readFastqcZips()}

\item{perc}{return percentage of sequences if TRUE, number of sequences otherwise}

\item{nreads}{return value from nreads.fastqc(), required if perc=TRUE}
}
\value{
numeric matrix with entries in rows and number (percentage) of sequences per quality in columns
}
\description{
Return the number, or percentage, of sequences per quality score per entry in a set of zip archives generated by FastQC.
To be used after readFastqcZips().
}
\author{
Timothee Flutre [cre,aut]
}

