% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractFastqcSeqLengths}
\alias{extractFastqcSeqLengths}
\title{Returns the sequence length distribution per entry in a set of zip archives generated by FastQC.}
\usage{
extractFastqcSeqLengths(all.qc)
}
\arguments{
\item{all.qc}{return value from readFastqcZips()}
}
\value{
numeric matrix with entries in rows and positions along sequences in columns
}
\description{
To be used after readFastqcZips().
}
\author{
Timothee Flutre [cre,aut], Nicolas Rode [ctb]
}

