% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plotFastqcContent}
\alias{plotFastqcContent}
\title{Plot a variable content (adapter or N) as percentage along the sequences per entry in a set of zip archives generated by FastQC.}
\usage{
plotFastqcContent(content, max.datasets.per.plot = 25, lowest.perc = NULL,
  highest.perc = NULL, ylab = "Content (\%)", main = "Quality control",
  legend.x = "topleft", legend.y = NULL, legend.cex = 1,
  add.2nd.yaxis = TRUE)
}
\arguments{
\item{content}{return value from extractFastqcAdpContents() or extractFastqcBaseNs()}

\item{max.datasets.per.plot}{max number of datasets on the same plot}

\item{lowest.perc}{lowest percentage of content for the y-axis}

\item{highest.perc}{highest percentage of content for the y-axis}

\item{ylab}{a title for the y axis}

\item{main}{an overall title for the plot}

\item{legend.x}{x coordinate to position the legend (no legend if NULL)}

\item{legend.y}{y coordinate to position the legend}

\item{legend.cex}{numeric character expansion factor for legend labels}

\item{add.2nd.yaxis}{add a 2nd y-axis on the right side of the plot}
}
\value{
nothing
}
\description{
To be used after extractFastqcAdpContents() or extractFastqcBaseNs().
}
\author{
Timothee Flutre [cre,aut], Nicolas Rode [ctb]
}

