% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotFastqcNbseqQual}
\alias{plotFastqcNbseqQual}
\title{Plot the number, or percentage, of sequences per quality score with one curve per dataset.}
\usage{
plotFastqcNbseqQual(qual, perc = FALSE, ylim = NULL,
  max.datasets.per.plot = 25, main = "Quality control",
  legend.x = "topleft", legend.y = NULL, legend.cex = 1,
  add.2nd.yaxis = TRUE)
}
\arguments{
\item{qual}{return value from extractFastqcQuals()}

\item{perc}{value of perc used when qual was generated by extractFastqcQuals()}

\item{ylim}{left and right limits of the y-axis (will be min and max of qual by default)}

\item{max.datasets.per.plot}{max number of datasets on the same plot}

\item{main}{an overall title for the plot}

\item{legend.x}{x coordinate to position the legend (no legend if NULL)}

\item{legend.y}{y coordinate to position the legend}

\item{legend.cex}{numeric character expansion factor for legend labels}

\item{add.2nd.yaxis}{add a 2nd y-axis on the right side of the plot}
}
\value{
nothing
}
\description{
To be used after extractFastqcQuals().
}
\author{
Timothee Flutre [cre,aut], Nicolas Rode [ctb]
}

