% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotFastqcNreads}
\alias{plotFastqcNreads}
\title{Plot FastQC}
\usage{
plotFastqcNreads(x, main = "", cex = 1)
}
\arguments{
\item{x}{return value from extractFastqcNreads()}

\item{main}{an overall title for the plot}

\item{cex}{numeric character expansion factor for x-axis labels}
}
\value{
nothing
}
\description{
Creates a bar plot with the number of sequences per entry in a set of zip archives generated by FastQC.
To be used after extractFastqcNreads().
}
\author{
Timothee Flutre [cre,aut]
}

