% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotFastqcSeqLengths}
\alias{plotFastqcSeqLengths}
\title{Plot the distribution of sequence lengths per entry in a set of zip archives generated by FastQC.}
\usage{
plotFastqcSeqLengths(seq.length, max.datasets.per.plot = 25,
  lowest.len = NULL, highest.len = NULL, main = "Quality control",
  ylab = "Number of sequences", legend.x = "topleft", legend.y = NULL,
  legend.cex = 1, add.2nd.yaxis = TRUE)
}
\arguments{
\item{seq.length}{return value from extractFastqcSeqLengths()}

\item{max.datasets.per.plot}{max number of datasets on the same plot}

\item{lowest.len}{lowest sequence length for the y-axis}

\item{highest.len}{highest sequence length for the y-axis}

\item{main}{an overall title for the plot}

\item{ylab}{label for the y-axis}

\item{legend.x}{x coordinate to position the legend (no legend if NULL)}

\item{legend.y}{y coordinate to position the legend}

\item{legend.cex}{numeric character expansion factor for legend labels}

\item{add.2nd.yaxis}{add a 2nd y-axis on the right side of the plot}
}
\value{
nothing
}
\description{
To be used after extractFastqcSeqLengths().
}
\author{
Timothee Flutre [cre,aut], Nicolas Rode [ctb]
}

