% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readFastqcZips}
\alias{readFastqcZips}
\title{Read several zip archives generated by FastQC.}
\usage{
readFastqcZips(path = ".", glob = "*_fastqc.zip", verbose = 0)
}
\arguments{
\item{path}{character vector of the path to the directory containing the zip archives (will be followed by "*_fastqc.zip")}

\item{glob}{character vector with wildcard(s) to find zip archives}

\item{verbose}{verbosity level}
}
\value{
list of lists (one per zip archive)
}
\description{
Zip archives are decompressed in a temporary directory which is removed afterwards.
}
\author{
Timothee Flutre [cre,aut], Nicolas Rode [ctb]
}

