% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.R
\name{extractFastqcAdpContents}
\alias{extractFastqcAdpContents}
\title{Returns the adapter content along the sequences per entry in a set of zip archives generated by FastQC.}
\usage{
extractFastqcAdpContents(all.qc, adp = "Illumina Universal Adapter")
}
\arguments{
\item{all.qc}{return value from readFastqcZips()}

\item{adp}{name of the adapter to plot (default="Illumina Universal Adapter")}
}
\value{
numeric matrix with entries in rows and positions along sequences in columns
}
\description{
To be used after readFastqcZips().
}
\author{
Timothee Flutre [cre,aut], Nicolas Rode [ctb]
}

