% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractFastqcAdpContents}
\alias{extractFastqcAdpContents}
\title{Extract FastQC}
\usage{
extractFastqcAdpContents(all.qc, adp = "Illumina Universal Adapter")
}
\arguments{
\item{all.qc}{return value from readFastqcZips()}

\item{adp}{name of the adapter to plot (default="Illumina Universal Adapter")}
}
\value{
numeric matrix with entries in rows and positions along sequences in columns
}
\description{
Returns the adapter content along the sequences per entry in a set of zip archives generated by FastQC.
To be used after readFastqcZips().
When reads are long, FastQC can average adapter content over a given range, e.g. over two successive bases, such as "11-12".
In such a case, because the ranges can be different for different fastq files in \code{all.qc}, this function will set adapter content to the first base of the range in the ouput, here "11", the other, here "12", will get NA.
}
\author{
Timothee Flutre [cre,aut], Nicolas Rode [ctb]
}

