% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractFastqcBaseNs}
\alias{extractFastqcBaseNs}
\title{Extract FastQC}
\usage{
extractFastqcBaseNs(all.qc)
}
\arguments{
\item{all.qc}{return value from readFastqcZips()}
}
\value{
numeric matrix with entries in rows and positions along sequences in columns
}
\description{
Returns the N counts per entry in a set of zip archives generated by FastQC.
To be used after readFastqcZips().
}
\author{
Timothee Flutre [cre,aut]
}

