% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readFastqcTxt}
\alias{readFastqcTxt}
\title{Read FastQC}
\usage{
readFastqcTxt(file)
}
\arguments{
\item{file}{the name of the file which the data are to be read from}
}
\value{
list
}
\description{
Read a "fastqc_data.txt" file generated by FastQC.
Heavily inspired from readFastQC() in the Repitools package.
}
\author{
Timothee Flutre [cre,aut]
}

